db and tables
